Backups will most likely not be managed with ansible
See https://www.postgresql.org/docs/9.1/static/backup.html on details
how to backup postgres. The article lists 3 approaches where the first
one fits best because the database must not be shutdown when using
`pg_dump` or `pg_dumpall` to generate a list of SQL statements which can
regenerate the database state at the time of backup. This does also work
when moving between computer architectures (32-bit/64-bit), filesystems
and distributions. Thus, this is to be preferred to a backup on
filesystem level.