Add a script to generate a valid self-signed certificate
This is based on this gist: https://gist.github.com/jdeathe/4c08460eb0bac99da32748bcbda39333

Unfortunately, since FF 48 and Chrome 58 it is required to set a
SubjectAltName (SAN) in the certificate because the CommonName (CN)
support was removed.

There is also this https://alexanderzeitler.com/articles/Fixing-Chrome-missing_subjectAltName-selfsigned-cert-openssl/
article which uses a custom root CA. Maybe I give this a try.

Nonetheless, I can't thank letsencrypt enough because actually they made
it easier to get a valid certificate then to generate a self-signed
one.