Install Python2 which ansible requires
This is achieved through a command of type `raw` that ansible simply
tells to run it as-is via SSH. With other words, raw commands only
require SSH to be working on a remote machine and are executed as
argument to `ssh <remote-machine> '<raw-cmd>'`.
Gathering facts must be deactivated because it would take place before
`pre_tasks` and fail since /usr/bin/python[2] is probably not available on the
machine.

The solution was shamlessly copied from https://github.com/maqpie/deploy-drone/blob/master/setup-server.yml#L13