Upgrade to prometheus 2.0.0
Old prometheus metrics data is incompatible with the new metrics
storage, the migration guide [1] recommends to run a non-scraping 1.8
instance besides the new prometheus and to configure it as remote_read.
Because my metric data is no that important I decided to create a backup
`sudo tar -C /home/prometheus/data -cJvf prometheus-data-backup.tar.xz .`
and remove it.

Command line flag format has changed and now uses double dashes as a
prefix and options for `--storage.{local,remote}` are gone.

[1]: https://prometheus.io/docs/prometheus/2.0/migration/